+++ data-migrations/2025-04-01-0738-drop-planVersions.sql
@@ -1,0 +1,2 @@
+# Drop the planVersions table
+DROP TABLE IF EXISTS planVersions CASCADE;