+++ src/controllers/__tests__/signinController.spec.ts
@@ -58,8 +58,6 @@ const mockedUser: UserModel.User = {
   hasErrors: jest.fn(),
 };
 
-jest.mock('../../models/User');
-
 describe('signinController', () => {
   let mockRequest: Partial<Request>;
   let mockResponse: Partial<Response>;