+++ src/resolvers/plan.ts
@@ -10,9 +10,8 @@ import { PlanContributor } from "../models/Contributor";
 import { PlanFunder } from "../models/Funder";
 import { Resolvers } from "../types";
 import { VersionedTemplate } from "../models/VersionedTemplate";
-import { createPlanVersion } from "../services/planService";
-import { getCurrentDate, randomHex } from "../utils/helpers";
 import { Answer } from "../models/Answer";
+import { ProjectCollaboratorAccessLevel } from "../models/Collaborator";
 
 export const resolvers: Resolvers = {
   Query: {
@@ -49,7 +48,7 @@ export const resolvers: Resolvers = {
         }
 
         const project = await Project.findById(reference, context, plan.projectId);
-        if (project && hasPermissionOnProject(context, project)) {
+        if (project && await hasPermissionOnProject(context, project)) {
           return plan;
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -69,29 +68,18 @@ export const resolvers: Resolvers = {
       try {
         if (isAuthorized(context.token)) {
           const project = await Project.findById(reference, context, projectId);
+          const versionedTemplate = await VersionedTemplate.findById(reference, context, versionedTemplateId);
 
           if (!project) {
             throw NotFoundError(`Project with ID ${projectId} not found`);
           }
+          if (!versionedTemplate) {
+            throw NotFoundError(`Template with ID ${versionedTemplateId} not found`);
+          }
 
           if (await hasPermissionOnProject(context, project)) {
             const plan = new Plan({ projectId, versionedTemplateId });
-            // TODO: Uncomment this bit once we do more thorough testing of comms with the DMPHub
-            //
-            // Send the plan to the DMPHub so that it stores it and assigns a DMP ID
-            // const dmp = await syncWithDMPHub(context, plan, reference);
-            // formatLogMessage(context).debug({ dmp }, `DMP from the DMPHub API: for plan: ${plan.id}`);
-
-            // Record the DMP ID that was assigned by the DMPHub
-            // plan.dmpId = dmp?.dmp_id?.identifier;
-            plan.dmpId = randomHex(16);
-
-            const newPlan = await plan.create(context);
-            if (newPlan) {
-              return new Plan(newPlan);
-            }
-
-            return null;
+            return plan.create(context);
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -104,34 +92,26 @@ export const resolvers: Resolvers = {
     },
 
     // Delete a plan
-    archivePlan: async (_, { dmpId }, context: MyContext): Promise<Plan> => {
+    archivePlan: async (_, { planId }, context: MyContext): Promise<Plan> => {
       const reference = 'archive plan resolver';
       try {
         if (isAuthorized(context.token)) {
-          const plan = await Plan.findByDMPId(reference, context, dmpId);
+          const plan = await Plan.findById(reference, context, planId);
           if (!plan) {
-            throw NotFoundError(`Plan with DMP ID ${dmpId} not found`);
+            throw NotFoundError(`Plan with id ${planId} not found`);
           }
+
           const project = await Project.findById(reference, context, plan.projectId);
+          if (plan.isPublished()) {
+            plan.addError('general', 'Plan is already published and cannot be archived');
+          }
+
           if (await hasPermissionOnProject(context, project)) {
-            // First create a version snapshot (before making changes)
-            const newVersion = createPlanVersion(context, plan, reference);
-            if (newVersion) {
-              plan.status = PlanStatus.ARCHIVED;
-              const deletedPlan = await plan.delete(context);
-
-              if (deletedPlan) {
-                // TODO: Uncomment this bit once we do more thorough testing of comms with the DMPHub
-                //
-                // asyncronously tombstone the DMP in the DMPHub (after making changes)
-                // syncWithDMPHub(context, plan, reference);
-
-                return deletedPlan;
-              }
+            if (!plan.hasErrors()) {
+              return await plan.delete(context);
+            } else {
+              return plan;
             }
-
-            plan.addError('general', 'Unable to archive plan. Failed to create a version snapshot');
-            return plan;
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -171,84 +151,30 @@ export const resolvers: Resolvers = {
     },
 
     // Publish/register the plan with the DOI registrar (e.g. EZID/DataCite)
-    publishPlan: async (_, { dmpId, visibility }, context: MyContext): Promise<Plan> => {
+    publishPlan: async (_, { planId, visibility = PlanVisibility.PRIVATE }, context: MyContext): Promise<Plan> => {
       const reference = 'publish plan resolver';
       try {
         if (isAuthorized(context.token)) {
-          const plan = await Plan.findByDMPId(reference, context, dmpId);
+          const plan = await Plan.findById(reference, context, planId);
           if (!plan) {
-            throw NotFoundError(`Plan with DMP ID ${dmpId} not found`);
+            throw NotFoundError(`Plan with id ${planId} not found`);
           }
-          const project = await Project.findById(reference, context, plan.projectId);
-          if (await hasPermissionOnProject(context, project)) {
-            if (project.isTestProject) {
-              plan.addError('general', 'Test projects cannot be published');
-            } else if (plan.registered) {
-              plan.addError('general', 'Plan is already published');
-            }
+          if (plan.isPublished()) {
+            plan.addError('general', 'Plan is already published');
+          }
 
+          const project = await Project.findById(reference, context, plan.projectId);
+          if (await hasPermissionOnProject(context, project, ProjectCollaboratorAccessLevel.OWN)) {
             if (!plan.hasErrors()) {
-              // First create a version snapshot (before making changes)
-              const newVersion = createPlanVersion(context, plan, reference);
-              if (newVersion) {
-                plan.status = PlanStatus.PUBLISHED;
-                plan.registered = new Date().toISOString();
-                plan.registeredById = context.token.id;
+              if (project.isTestProject) {
+                plan.addError('general', 'Test projects cannot be published');
+              } else if (plan.isPublished()) {
+                plan.addError('general', 'Plan is already published');
+              }
+
+              if (!plan.hasErrors()) {
                 plan.visibility = visibility as PlanVisibility;
-
-                const publishedPlan = await plan.update(context);
-                if (publishedPlan) {
-                  // TODO: Uncomment this bit once we do more thorough testing of comms with the DMPHub
-                  //
-                  // asyncronously tombstone the DMP in the DMPHub (after making changes)
-                  // syncWithDMPHub(context, plan, reference);
-
-                  return publishedPlan;
-                }
-              }
-            }
-
-            return plan;
-          }
-        }
-        throw context?.token ? ForbiddenError() : AuthenticationError();
-      } catch (err) {
-        if (err instanceof GraphQLError) throw err;
-
-        formatLogMessage(context).error(err, `Failure in ${reference}`);
-        throw InternalServerError();
-      }
-    },
-
-    // Mark the plan as complete
-    markPlanAsComplete: async (_, { dmpId }, context: MyContext): Promise<Plan> => {
-      const reference = 'mark plan complete resolver';
-      try {
-        if (isAuthorized(context.token)) {
-          const plan = await Plan.findByDMPId(reference, context, dmpId);
-          if (!plan) {
-            throw NotFoundError(`Plan with DMP ID ${dmpId} not found`);
-          }
-          const project = await Project.findById(reference, context, plan.projectId);
-          if (await hasPermissionOnProject(context, project)) {
-            if (plan.registered) {
-              plan.addError('general', 'Plan is already published');
-            } else {
-              // First create a version snapshot (before making changes)
-              const newVersion = createPlanVersion(context, plan, reference);
-              if (newVersion) {
-                plan.status = PlanStatus.COMPLETE;
-                plan.lastSynced = getCurrentDate();
-
-                const updatedPlan = await plan.update(context);
-                if (updatedPlan) {
-                  // TODO: Uncomment this bit once we do more thorough testing of comms with the DMPHub
-                  //
-                  // asyncronously tombstone the DMP in the DMPHub (after making changes)
-                  // syncWithDMPHub(context, plan, reference);
-
-                  return updatedPlan;
-                }
+                return await plan.publish(context);
               }
             }
             return plan;
@@ -263,38 +189,18 @@ export const resolvers: Resolvers = {
       }
     },
 
-    // Mark the plan as a draft
-    markPlanAsDraft: async (_, { dmpId }, context: MyContext): Promise<Plan> => {
-      const reference = 'mark plan draft resolver';
+    updatePlanStatus: async (_, { planId, status }, context: MyContext): Promise<Plan> => {
+      const reference = 'update plan status resolver';
       try {
         if (isAuthorized(context.token)) {
-          const plan = await Plan.findByDMPId(reference, context, dmpId);
+          const plan = await Plan.findById(reference, context, planId);
           if (!plan) {
-            throw NotFoundError(`Plan with DMP ID ${dmpId} not found`);
+            throw NotFoundError(`Plan with id ${planId} not found`);
           }
           const project = await Project.findById(reference, context, plan.projectId);
-          if (await hasPermissionOnProject(context, project)) {
-            if (plan.registered) {
-              plan.addError('general', 'Plan is already published');
-            } else {
-              // First create a version snapshot (before making changes)
-              const newVersion = createPlanVersion(context, plan, reference);
-              if (newVersion) {
-                plan.status = PlanStatus.DRAFT;
-                plan.lastSynced = getCurrentDate();
-
-                const updatedPlan = await plan.update(context);
-                if (updatedPlan) {
-                  // TODO: Uncomment this bit once we do more thorough testing of comms with the DMPHub
-                  //
-                  // asyncronously tombstone the DMP in the DMPHub (after making changes)
-                  // syncWithDMPHub(context, plan, reference);
-
-                  return updatedPlan;
-                }
-              }
-            }
-            return plan;
+          if (await hasPermissionOnProject(context, project, ProjectCollaboratorAccessLevel.OWN)) {
+            plan.status = status as PlanStatus;
+            return await plan.update(context);
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();