+++ src/resolvers/relatedWork.ts
@@ -7,6 +7,8 @@ import { RelatedWork } from '../models/RelatedWork';
 import { GraphQLError } from 'graphql';
 import { Project } from '../models/Project';
 import { hasPermissionOnProject } from '../services/projectService';
+import { Plan } from '../models/Plan';
+import { addVersion } from '../models/PlanVersion';
 
 export const resolvers: Resolvers = {
   Query: {
@@ -16,7 +18,7 @@ export const resolvers: Resolvers = {
       try {
         if (isAuthorized(context.token)) {
           const project = await Project.findById(reference, context, projectId);
-          if (project && hasPermissionOnProject(context, project)) {
+          if (project && await hasPermissionOnProject(context, project)) {
             return await RelatedWork.findByProjectId(reference, context, projectId);
           }
         }
@@ -35,7 +37,7 @@ export const resolvers: Resolvers = {
         if (isAuthorized(context.token)) {
           const relatedWork = await RelatedWork.findById(reference, context, id);
           const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && hasPermissionOnProject(context, project)) {
+          if (project && await hasPermissionOnProject(context, project)) {
             return relatedWork;
           }
         }
@@ -56,12 +58,18 @@ export const resolvers: Resolvers = {
       try {
         if (isAuthorized(context.token)) {
           const project = await Project.findById(reference, context, input.projectId);
-          if (project && hasPermissionOnProject(context, project)) {
+          if (project && await hasPermissionOnProject(context, project)) {
             const relatedWork = new RelatedWork(input);
+            const newRelatedWork = await relatedWork.create(context);
 
-            // TODO: We need to generate the plan version snapshot and sync with DMPHub for each plan
-
-            return await relatedWork.create(context);
+            if (newRelatedWork && !newRelatedWork.hasErrors()) {
+              // Version all of the plans (if any) and sync with the DMPHub
+              const plans = await Plan.findByProjectId(reference, context, project.id);
+              for (const plan of plans) {
+                await addVersion(context, plan, reference);
+              }
+            }
+            return newRelatedWork;
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -84,12 +92,19 @@ export const resolvers: Resolvers = {
           }
 
           const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && hasPermissionOnProject(context, project)) {
+          if (project && await hasPermissionOnProject(context, project)) {
             const toUpdate = new RelatedWork({ ...relatedWork, ...input });
+            const updated = await toUpdate.update(context);
 
-            // TODO: We need to generate the plan version snapshot and sync with DMPHub for each plan
+            if(updated && !updated.hasErrors()) {
+              // Version all of the plans (if any) and sync with the DMPHub
+              const plans = await Plan.findByProjectId(reference, context, project.id);
+              for (const plan of plans) {
+                await addVersion(context, plan, reference);
+              }
+            }
 
-            return await toUpdate.update(context);
+            return updated;
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -112,11 +127,18 @@ export const resolvers: Resolvers = {
           }
 
           const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && hasPermissionOnProject(context, project)) {
+          if (project && await hasPermissionOnProject(context, project)) {
+            const removed = await relatedWork.delete(context);
 
-            // TODO: We need to generate the plan version snapshot and sync with DMPHub for each plan
+            if(removed && !removed.hasErrors()) {
+              // Version all of the plans (if any) and sync with the DMPHub
+              const plans = await Plan.findByProjectId(reference, context, project.id);
+              for (const plan of plans) {
+                await addVersion(context, plan, reference);
+              }
+            }
 
-            return await relatedWork.delete(context);
+            return removed;
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();