+++ src/services/__tests__/templateService.spec.ts
@@ -6,13 +6,13 @@ import { TemplateCollaborator } from '../../models/Collaborator';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { isSuperAdmin } from '../authService';
 import { logger } from '../../__mocks__/logger';
-import { MySQLDataSource } from '../../datasources/mySQLDataSource';
+import { mysql } from '../../datasources/mysql';
 import { buildContext, mockToken } from '../../__mocks__/context';
 import { Section } from '../../models/Section';
 import { getRandomEnumValue } from '../../__tests__/helpers';
 import { getCurrentDate } from '../../utils/helpers';
 
-// Pulling context in here so that the MySQLDataSource gets mocked
+// Pulling context in here so that the mysql gets mocked
 jest.mock('../../context.ts');
 
 let context;
@@ -29,7 +29,6 @@ afterEach(() => {
 
 describe('hasPermissionOnTemplate', () => {
   let template;
-  let mockQuery;
   let mockIsSuperAdmin;
   let mockFindByTemplateAndEmail;
   let context;
@@ -38,13 +37,11 @@ describe('hasPermissionOnTemplate', () => {
     jest.resetAllMocks();
 
     // Cast getInstance to a jest.Mock type to use mockReturnValue
-    (MySQLDataSource.getInstance as jest.Mock).mockReturnValue({
+    (mysql.getInstance as jest.Mock).mockReturnValue({
       query: jest.fn(), // Initialize the query mock function here
     });
 
-    const instance = MySQLDataSource.getInstance();
-    mockQuery = instance.query as jest.MockedFunction<typeof instance.query>;
-    context = { logger, dataSources: { sqlDataSource: { query: mockQuery } } };
+    context = buildContext(logger, mockToken());
 
     mockIsSuperAdmin = jest.fn();
     (isSuperAdmin as jest.Mock) = mockIsSuperAdmin;