+++ src/services/sectionService.ts
@@ -106,6 +106,8 @@ export const cloneSection = (
 
 // Determine whether the specified user has permission to access the Section
 export const hasPermissionOnSection = async (context: MyContext, templateId: number): Promise<boolean> => {
+  if (!context || !context.token) return false;
+
   // Find associated template info
   const template = await Template.findById('section resolver.hasPermission', context, templateId);